third_party/blink/web_tests/fast/forms/interactive-validation-formnovalidate-expected.txt

Test if the form is submitted with a submit button with formnovalidate.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS The form should not be submitted.
PASS location.search.indexOf("i0=") != -1 is true
TEST COMPLETE
 
